pyproject.toml
[project]
name = "hireflow"
version = "0.1.0"
description = "Local-first job application tracker powered by Gmail and Claude AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "google-api-python-client>=2.195.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]
